Move painting to src root
