+++ lib/src/main/java/com/zetcode/painting/icesid.jpg
