+++ lib/src/main/java/com/zetcode/painting/java.png
