+++ lib/src/main/java/com/zetcode/painting/pane.png
