+++ lib/src/main/java/com/zetcode/painting/slate.png
